# default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "API documentation for my project",
    "version": "1.0.0",
    "title": "My API Documentation",
    "contact": {
      "name": "John Doe",
      "url": "http://example.com",
      "email": "john@example.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/admin/meter/create": {
      "post": {
        "tags": [
          "电表管理"
        ],
        "summary": "createMeter",
        "operationId": "createMeterUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "meterCreateDTO",
            "description": "meterCreateDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MeterCreateDTO",
              "originalRef": "MeterCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«object»",
              "originalRef": "R«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/meter/delete/{meterId}": {
      "put": {
        "tags": [
          "电表管理"
        ],
        "summary": "deleteMeter",
        "operationId": "deleteMeterUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "meterId",
            "in": "path",
            "description": "meterId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«object»",
              "originalRef": "R«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/meter/edit": {
      "post": {
        "tags": [
          "电表管理"
        ],
        "summary": "editMeter",
        "operationId": "editMeterUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "meterBindDTO",
            "description": "meterBindDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MeterEditDTO",
              "originalRef": "MeterEditDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«object»",
              "originalRef": "R«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/meter/inspection/{meterId}": {
      "post": {
        "tags": [
          "电表管理"
        ],
        "summary": "inspectionCreate",
        "operationId": "inspectionCreateUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "meterInspectionDTO",
            "description": "meterInspectionDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MeterInspectionDTO",
              "originalRef": "MeterInspectionDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«object»",
              "originalRef": "R«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/meter/model": {
      "get": {
        "tags": [
          "电表管理"
        ],
        "summary": "getMeterModel",
        "operationId": "getMeterModelUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/meter/page": {
      "get": {
        "tags": [
          "电表管理"
        ],
        "summary": "queryMeterPage",
        "operationId": "queryMeterPageUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "meterId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«MeterPageVO»",
              "originalRef": "PageDTO«MeterPageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/meter/{meterId}": {
      "get": {
        "tags": [
          "电表管理"
        ],
        "summary": "getMeterDetail",
        "operationId": "getMeterDetailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "meterId",
            "in": "path",
            "description": "meterId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MeterDetailVO",
              "originalRef": "MeterDetailVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/login": {
      "get": {
        "tags": [
          "登录管理"
        ],
        "summary": "getPublicKey",
        "operationId": "getPublicKeyUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/login/captcha": {
      "get": {
        "tags": [
          "登录管理"
        ],
        "summary": "验证码",
        "description": "验证码",
        "operationId": "captchaUsingGET",
        "produces": [
          "image/png"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/login/login": {
      "post": {
        "tags": [
          "登录管理"
        ],
        "summary": "login",
        "operationId": "loginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "adminFormDTO",
            "description": "adminFormDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminFormDTO",
              "originalRef": "AdminFormDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«LoginVO»",
              "originalRef": "R«LoginVO»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/login/logout": {
      "post": {
        "tags": [
          "登录管理"
        ],
        "summary": "logout",
        "operationId": "logoutUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/login/refresh": {
      "get": {
        "tags": [
          "登录管理"
        ],
        "summary": "刷新token",
        "operationId": "refreshTokenUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "adminToken",
            "in": "query",
            "description": "adminToken",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AdminDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null,
          "example": 1001
        },
        "rememberMe": {
          "type": "boolean",
          "refType": null,
          "example": true
        },
        "roleName": {
          "type": "string",
          "refType": null,
          "example": "系统管理员"
        },
        "userName": {
          "type": "string",
          "refType": null,
          "example": "admin"
        }
      },
      "title": "AdminDTO"
    },
    "AdminFormDTO": {
      "type": "object",
      "required": [
        "account",
        "code",
        "key",
        "password"
      ],
      "properties": {
        "account": {
          "type": "string",
          "refType": null,
          "example": "admin"
        },
        "code": {
          "type": "string",
          "refType": null,
          "example": "A6F2"
        },
        "key": {
          "type": "string",
          "refType": null,
          "example": "a1b2c3d4e5f6"
        },
        "password": {
          "type": "string",
          "refType": null,
          "example": "9d4e1e23bd5b727046a9e3b4b7db57bd8d6ee684"
        },
        "rememberMe": {
          "type": "boolean",
          "refType": null,
          "example": true
        }
      },
      "title": "AdminFormDTO"
    },
...
    "LoginVO": {
      "type": "object",
      "properties": {
        "adminDTO": {
          "$ref": "#/definitions/AdminDTO",
          "originalRef": "AdminDTO",
          "refType": "AdminDTO"
        },
        "token": {
          "type": "string",
          "refType": null
        }
      },
      "title": "LoginVO"
    },
    "MeterCreateDTO": {
      "type": "object",
      "properties": {
        "installDate": {
          "type": "string",
          "format": "date-time",
          "refType": null,
          "example": "2023-03-15T09:30:00"
        },
        "installPlace": {
          "type": "string",
          "refType": null,
          "example": "北京市海淀区中关村南路10号3单元501室"
        },
        "model": {
          "type": "string",
          "refType": null,
          "example": "DDS-102"
        },
        "status": {
          "type": "string",
          "refType": null,
          "example": "正常"
        }
      },
      "title": "MeterCreateDTO"
    },
    "MeterDetailVO": {
      "type": "object",
      "properties": {
        "endingReading": {
          "type": "number",
          "example": 2458.75
        },
        "faultDescription": {
          "type": "string",
          "example": "显示屏偶尔闪烁"
        },
        "inspectionResult": {
          "type": "string",
          "example": "正常，误差在允许范围内"
        },
        "inspectionStatus": {
          "type": "string",
          "example": "已完成"
        },
        "inspectionTime": {
          "type": "string",
          "format": "date-time",
          "example": "2023-04-20T14:30:00"
        },
        "inspectionType": {
          "type": "string",
          "example": "季度检查"
        },
        "inspectorName": {
          "type": "string",
          "example": "王工"
        },
        "installDate": {
          "type": "string",
          "format": "date",
          "example": "2022-11-15"
        },
        "installPlace": {
          "type": "string",
          "example": "北京市海淀区中关村南路10号3单元501室"
        },
        "lastMeterReadingDate": {
          "type": "string",
          "format": "date-time",
          "example": "2023-05-01T10:15:00"
        },
        "model": {
          "type": "string",
          "example": "DDS-102"
        },
        "nextInspectionTime": {
          "type": "string",
          "format": "date-time",
          "example": "2023-07-20T14:30:00"
        },
        "remark": {
          "type": "string",
          "example": "用户反映电表读数变化过快，需下次重点检查"
        },
        "repairCost": {
          "type": "number",
          "example": 150.00
        },
        "solution": {
          "type": "string",
          "example": "更换显示屏驱动芯片"
        },
        "startMeterReadingDate": {
          "type": "string",
          "format": "date-time",
          "example": "2023-04-01T10:15:00"
        },
        "startReading": {
          "type": "number",
          "example": 2300.50
        },
        "status": {
          "type": "string",
          "example": "正常"
        },
        "userName": {
          "type": "string",
          "example": "张小明"
        }
      },
      "title": "MeterDetailVO"
    },
    "MeterEditDTO": {
      "type": "object",
      "required": [
        "meterId"
      ],
      "properties": {
        "idCardNo": {
          "type": "string",
          "refType": null,
          "example": "110101199001011234"
        },
        "installDate": {
          "type": "string",
          "format": "date-time",
          "refType": null,
          "example": "2023-03-15T09:30:00"
        },
        "installPlace": {
          "type": "string",
          "refType": null,
          "example": "北京市海淀区中关村南路10号3单元501室"
        },
        "meterId": {
          "type": "integer",
          "format": "int64",
          "refType": null,
          "example": 20230001
        },
        "model": {
          "type": "string",
          "refType": null,
          "example": "DDS-102"
        },
        "status": {
          "type": "string",
          "refType": null,
          "example": "正常"
        }
      },
      "title": "MeterEditDTO"
    },
    "MeterInspectionDTO": {
      "type": "object",
      "properties": {
        "faultDescription": {
          "type": "string",
          "refType": null,
          "example": "读数跳动过快，显示屏偶尔闪烁"
        },
        "inspectionResult": {
          "type": "string",
          "refType": null,
          "example": "需要维修"
        },
        "inspectionStatus": {
          "type": "string",
          "refType": null,
          "example": "已完成"
        },
        "inspectionTime": {
          "type": "string",
          "format": "date-time",
          "refType": null,
          "example": "2023-06-10T13:45:00"
        },
        "inspectionType": {
          "type": "string",
          "refType": null,
          "example": "故障检查"
        },
        "inspectorName": {
          "type": "string",
          "refType": null,
          "example": "李维修"
        },
        "meterId": {
          "type": "string",
          "refType": null,
          "example": "20230001"
        },
        "nextInspectionTime": {
          "type": "string",
          "format": "date-time",
          "refType": null,
          "example": "2023-09-10T13:45:00"
        },
        "remark": {
          "type": "string",
          "refType": null,
          "example": "用户反映电费增长异常，需查明原因"
        },
        "repairCost": {
          "type": "number",
          "refType": null,
          "example": 180.50
        },
        "solution": {
          "type": "string",
          "refType": null,
          "example": "更换主控板和显示驱动电路"
        }
      },
      "title": "MeterInspectionDTO"
    },
    "MeterPageVO": {
      "type": "object",
      "properties": {
        "endingReading": {
          "type": "number",
          "refType": null,
          "example": 3250.75
        },
        "id": {
          "type": "string",
          "refType": null,
          "example": "20230001"
        },
        "installDate": {
          "type": "string",
          "format": "date",
          "refType": null,
          "example": "2022-11-15"
        },
        "installPlace": {
          "type": "string",
          "refType": null,
          "example": "北京市海淀区中关村南路10号3单元501室"
        },
        "lastMeterReadingDate": {
          "type": "string",
          "format": "date-time",
          "refType": null,
          "example": "2023-06-01T10:30:00"
        },
        "model": {
          "type": "string",
          "refType": null,
          "example": "DDS-102"
        },
        "startReading": {
          "type": "number",
          "refType": null,
          "example": 3050.25
        },
        "status": {
          "type": "string",
          "refType": null,
          "example": "正常"
        },
        "userName": {
          "type": "string",
          "refType": null,
          "example": "张小明"
        }
      },
      "title": "MeterPageVO"
    },
...
    "PageDTO«MeterPageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MeterPageVO",
            "originalRef": "MeterPageVO"
          }
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PageDTO«MeterPageVO»"
    },
...
    "R«LoginVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "业务状态码，200-成功，其它-失败",
          "example": 200
        },
        "data": {
          "description": "响应数据",
          "$ref": "#/definitions/LoginVO",
          "originalRef": "LoginVO",
          "example": {
            "adminDTO": {
              "id": 1001,
              "rememberMe": true,
              "roleName": "系统管理员",
              "userName": "admin"
            },
            "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6ImFkbWluIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          }
        },
        "msg": {
          "type": "string",
          "example": "登录成功",
          "description": "响应消息"
        },
        "requestId": {
          "type": "string",
          "example": "1af123c11412e",
          "description": "请求id"
        }
      },
      "title": "R«LoginVO»",
      "description": "通用响应结果"
    },
    "R«object»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "业务状态码，200-成功，其它-失败",
          "example": 200
        },
        "data": {
          "type": "object",
          "description": "响应数据",
          "example": null
        },
        "msg": {
          "type": "string",
          "example": "操作成功",
          "description": "响应消息"
        },
        "requestId": {
          "type": "string",
          "example": "1af123c11412e",
          "description": "请求id"
        }
      },
      "title": "R«object»",
      "description": "通用响应结果"
    },
...
  }
}